third_party/WebKit/LayoutTests/crypto/hmac-generateKey-parseAlgorithm-failures.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests generateKey() with bad HMAC parameters.");
jsTestIsAsync = true;

extractable = true;
keyUsages = ['sign', 'verify'];

Promise.resolve(null).then(function() {
    return crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: -3}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'hmac', hash: {name: ''}, length: 48}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: 5000000000}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: NaN}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: Infinity}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: -Infinity}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: crypto}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: undefined}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: true}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);
}).then(finishJSTest, failAndFinishJSTest);

</script>

</body>
</html>
